requirements.txt
aiohappyeyeballs>=2.6
aiohttp>=3.13
aiosignal>=1.4
anyio>=4.13
asgiref>=3.11
attrs>=26.1
certifi>=2026.2
channels>=4.3,<5.0
charset-normalizer>=3.4
click>=8.3
discord>=2.3
discord.py>=2.7
Django>=6.0,<7.0
djangorestframework>=3.17,<4.0
dotenv>=0.9
frozenlist>=1.8
h11>=0.16
httptools>=0.7
idna>=3.11
multidict>=6.7
propcache>=0.4
psycopg2-binary>=2.9
python-dotenv>=1.2
PyYAML>=6.0
requests>=2.33
sqlparse>=0.5
typing_extensions>=4.15
urllib3>=2.6
uvicorn>=0.43
uvloop>=0.22
watchfiles>=1.1
websockets>=16.0
yarl>=1.23